// src/app/model/employees.ts
/**
 * Employee: Dùng để chứa dữ liệu của nhân viên.
 * @author hathang
 */
export interface Employee{
    employeeId: number;
    employeeName: String;
    employeeBirthDate: String;
    departmentName: String;
    employeeEmail: String;
    employeeTelephone: String;
    certificationName: String;
    endDate: String;
    score: String;
}